Simplify: extract recommendations api group
